requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
Flask==3.1.1
Flask-Caching==2.3.1
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
ms-identity-python @ git+https://github.com/azure-samples/ms-identity-python@834583646df315b79bc89f54808f4ae9c2a1ce59
msal==1.32.3
msgspec==0.19.0
pycparser==2.22
PyJWT==2.10.1
redis==6.2.0
requests==2.32.4
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
waitress==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
